Refactor turn and game management
